+++ contracts/interfaces/liquidity/IStabilityPool.sol
@@ -6,19 +6,22 @@ pragma solidity ^0.7.0;
  * The Stability Pool holds LUSD tokens deposited by Stability Pool depositors.
  *
  * When a trove is liquidated, then depending on system conditions, some of its LUSD debt gets offset with
- * LUSD in the Stability Pool:  that is, the offset debt evaporates, and an equal amount of LUSD tokens in the Stability Pool is burned.
+ * LUSD in the Stability Pool:  that is, the offset debt evaporates, and an equal amount of LUSD tokens in the
+ * Stability Pool is burned.
  *
- * Thus, a liquidation causes each depositor to receive a LUSD loss, in proportion to their deposit as a share of total deposits.
- * They also receive an ETH gain, as the ETH collateral of the liquidated trove is distributed among Stability depositors,
- * in the same proportion.
+ * Thus, a liquidation causes each depositor to receive a LUSD loss, in proportion to their deposit as a share of total
+ * deposits.
+ * They also receive an ETH gain, as the ETH collateral of the liquidated trove is distributed among Stability
+ * depositors, in the same proportion.
  *
- * When a liquidation occurs, it depletes every deposit by the same fraction: for example, a liquidation that depletes 40%
- * of the total LUSD in the Stability Pool, depletes 40% of each deposit.
+ * When a liquidation occurs, it depletes every deposit by the same fraction: for example, a liquidation that
+ * depletes 40% of the total LUSD in the Stability Pool, depletes 40% of each deposit.
  *
- * A deposit that has experienced a series of liquidations is termed a "compounded deposit": each liquidation depletes the deposit,
- * multiplying it by some factor in range ]0,1[
+ * A deposit that has experienced a series of liquidations is termed a "compounded deposit": each liquidation depletes
+ * the deposit, multiplying it by some factor in range ]0,1[
  *
- * Please see the implementation spec in the proof document, which closely follows on from the compounded deposit / ETH gain derivations:
+ * Please see the implementation spec in the proof document, which closely follows on from the compounded deposit / ETH
+ * gain derivations:
  * https://github.com/liquity/liquity/blob/master/papers/Scalable_Reward_Distribution_with_Compounding_Stakes.pdf
  *
  * --- LQTY ISSUANCE TO STABILITY POOL DEPOSITORS ---
@@ -27,18 +30,18 @@ pragma solidity ^0.7.0;
  *
  * Each deposit is tagged with the address of the front end through which it was made.
  *
- * All deposits earn a share of the issued LQTY in proportion to the deposit as a share of total deposits. The LQTY earned
- * by a given deposit, is split between the depositor and the front end through which the deposit was made, based on the front end's kickbackRate.
+ * All deposits earn a share of the issued LQTY in proportion to the deposit as a share of total deposits. The LQTY
+ * earned by a given deposit, is split between the depositor and the front end through which the deposit was made,
+ * based on the front end's kickbackRate.
  *
  * Please see the system Readme for an overview:
  * https://github.com/liquity/dev/blob/main/README.md#lqty-issuance-to-stability-providers
  */
 interface IStabilityPool {
-
   // --- Events ---
 
-  event StabilityPoolETHBalanceUpdated(uint _newBalance);
-  event StabilityPoolLUSDBalanceUpdated(uint _newBalance);
+  event StabilityPoolETHBalanceUpdated(uint256 _newBalance);
+  event StabilityPoolLUSDBalanceUpdated(uint256 _newBalance);
 
   event BorrowerOperationsAddressChanged(address _newBorrowerOperationsAddress);
   event TroveManagerAddressChanged(address _newTroveManagerAddress);
@@ -49,24 +52,24 @@ interface IStabilityPool {
   event PriceFeedAddressChanged(address _newPriceFeedAddress);
   event CommunityIssuanceAddressChanged(address _newCommunityIssuanceAddress);
 
-  event P_Updated(uint _P);
-  event S_Updated(uint _S, uint128 _epoch, uint128 _scale);
-  event G_Updated(uint _G, uint128 _epoch, uint128 _scale);
+  event P_Updated(uint256 _P);
+  event S_Updated(uint256 _S, uint128 _epoch, uint128 _scale);
+  event G_Updated(uint256 _G, uint128 _epoch, uint128 _scale);
   event EpochUpdated(uint128 _currentEpoch);
   event ScaleUpdated(uint128 _currentScale);
 
-  event FrontEndRegistered(address indexed _frontEnd, uint _kickbackRate);
+  event FrontEndRegistered(address indexed _frontEnd, uint256 _kickbackRate);
   event FrontEndTagSet(address indexed _depositor, address indexed _frontEnd);
 
-  event DepositSnapshotUpdated(address indexed _depositor, uint _P, uint _S, uint _G);
-  event FrontEndSnapshotUpdated(address indexed _frontEnd, uint _P, uint _G);
-  event UserDepositChanged(address indexed _depositor, uint _newDeposit);
-  event FrontEndStakeChanged(address indexed _frontEnd, uint _newFrontEndStake, address _depositor);
+  event DepositSnapshotUpdated(address indexed _depositor, uint256 _P, uint256 _S, uint256 _G);
+  event FrontEndSnapshotUpdated(address indexed _frontEnd, uint256 _P, uint256 _G);
+  event UserDepositChanged(address indexed _depositor, uint256 _newDeposit);
+  event FrontEndStakeChanged(address indexed _frontEnd, uint256 _newFrontEndStake, address _depositor);
 
-  event ETHGainWithdrawn(address indexed _depositor, uint _ETH, uint _LUSDLoss);
-  event LQTYPaidToDepositor(address indexed _depositor, uint _LQTY);
-  event LQTYPaidToFrontEnd(address indexed _frontEnd, uint _LQTY);
-  event EtherSent(address _to, uint _amount);
+  event ETHGainWithdrawn(address indexed _depositor, uint256 _ETH, uint256 _LUSDLoss);
+  event LQTYPaidToDepositor(address indexed _depositor, uint256 _LQTY);
+  event LQTYPaidToFrontEnd(address indexed _frontEnd, uint256 _LQTY);
+  event EtherSent(address _to, uint256 _amount);
 
   // --- Functions ---
 
@@ -90,20 +93,22 @@ interface IStabilityPool {
    * - Sender is not a registered frontend
    * - _amount is not zero
    * ---
-   * - Triggers a LQTY issuance, based on time passed since the last issuance. The LQTY issuance is shared between *all* depositors and front ends
+   * - Triggers a LQTY issuance, based on time passed since the last issuance.
+   * The LQTY issuance is shared between *all* depositors and front ends
    * - Tags the deposit with the provided front end tag param, if it's a new deposit
    * - Sends depositor's accumulated gains (LQTY, ETH) to depositor
    * - Sends the tagged front end's accumulated LQTY gains to the tagged front end
    * - Increases deposit and tagged front end's stake, and takes new snapshots for each.
    */
-  function provideToSP(uint _amount, address _frontEndTag) external;
+  function provideToSP(uint256 _amount, address _frontEndTag) external;
 
   /*
    * Initial checks:
    * - _amount is zero or there are no under collateralized troves left in the system
    * - User has a non zero deposit
    * ---
-   * - Triggers a LQTY issuance, based on time passed since the last issuance. The LQTY issuance is shared between *all* depositors and front ends
+   * - Triggers a LQTY issuance, based on time passed since the last issuance.
+   * The LQTY issuance is shared between *all* depositors and front ends
    * - Removes the deposit's front end tag if it is a full withdrawal
    * - Sends all depositor's accumulated gains (LQTY, ETH) to depositor
    * - Sends the tagged front end's accumulated LQTY gains to the tagged front end
@@ -111,7 +116,7 @@ interface IStabilityPool {
    *
    * If _amount > userDeposit, the user withdraws all of their compounded deposit.
    */
-  function withdrawFromSP(uint _amount) external;
+  function withdrawFromSP(uint256 _amount) external;
 
   /*
    * Initial checks:
@@ -119,7 +124,8 @@ interface IStabilityPool {
    * - User has an open trove
    * - User has some ETH gain
    * ---
-   * - Triggers a LQTY issuance, based on time passed since the last issuance. The LQTY issuance is shared between *all* depositors and front ends
+   * - Triggers a LQTY issuance, based on time passed since the last issuance.
+   * The LQTY issuance is shared between *all* depositors and front ends
    * - Sends all depositor's LQTY gain to  depositor
    * - Sends all tagged front end's LQTY gain to the tagged front end
    * - Transfers the depositor's entire ETH gain from the Stability Pool to the caller's trove
@@ -136,7 +142,7 @@ interface IStabilityPool {
    * ---
    * Front end makes a one-time selection of kickback rate upon registering
    */
-  function registerFrontEnd(uint _kickbackRate) external;
+  function registerFrontEnd(uint256 _kickbackRate) external;
 
   /*
    * Initial checks:
@@ -146,23 +152,23 @@ interface IStabilityPool {
    * and transfers the Trove's ETH collateral from ActivePool to StabilityPool.
    * Only called by liquidation functions in the TroveManager.
    */
-  function offset(uint _debt, uint _coll) external;
+  function offset(uint256 _debt, uint256 _coll) external;
 
   /*
    * Returns the total amount of ETH held by the pool, accounted in an internal variable instead of `balance`,
    * to exclude edge cases like ETH received from a self-destruct.
    */
-  function getETH() external view returns (uint);
+  function getETH() external view returns (uint256);
 
   /*
    * Returns LUSD held in the pool. Changes when users deposit/withdraw, and when Trove debt is offset.
    */
-  function getTotalLUSDDeposits() external view returns (uint);
+  function getTotalLUSDDeposits() external view returns (uint256);
 
   /*
    * Calculates the ETH gain earned by the deposit since its last snapshots were taken.
    */
-  function getDepositorETHGain(address _depositor) external view returns (uint);
+  function getDepositorETHGain(address _depositor) external view returns (uint256);
 
   /*
    * Calculate the LQTY gain earned by a deposit since its last snapshots were taken.
@@ -170,24 +176,24 @@ interface IStabilityPool {
    * Otherwise, their cut of the deposit's earnings is equal to the kickbackRate, set by the front end through
    * which they made their deposit.
    */
-  function getDepositorLQTYGain(address _depositor) external view returns (uint);
+  function getDepositorLQTYGain(address _depositor) external view returns (uint256);
 
   /*
    * Return the LQTY gain earned by the front end.
    */
-  function getFrontEndLQTYGain(address _frontEnd) external view returns (uint);
+  function getFrontEndLQTYGain(address _frontEnd) external view returns (uint256);
 
   /*
    * Return the user's compounded deposit.
    */
-  function getCompoundedLUSDDeposit(address _depositor) external view returns (uint);
+  function getCompoundedLUSDDeposit(address _depositor) external view returns (uint256);
 
   /*
    * Return the front end's compounded stake.
    *
    * The front end's compounded stake is equal to the sum of its depositors' compounded deposits.
    */
-  function getCompoundedFrontEndStake(address _frontEnd) external view returns (uint);
+  function getCompoundedFrontEndStake(address _frontEnd) external view returns (uint256);
 
   /*
    * Fallback function
